feat: set id and description for the VIP-375
